% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_geneTF.R
\name{filter_geneTF}
\alias{filter_geneTF}
\title{Filter geneTF prior based on the gene list user provided}
\usage{
filter_geneTF(
  geneTF_prior,
  gene_list,
  TF_upbound = 0.8,
  TF_lowbound = 10,
  gene_lowbound = 5
)
}
\arguments{
\item{geneTF_prior}{geneTF prior}

\item{gene_list}{gene list user provided}

\item{TF_upbound}{the upper bound of percentage of TF targeted genes over all genes. The default is 80\%}

\item{TF_lowbound}{the lower bound of number of TF targeted genes. The defalut is 10}

\item{gene_lowbound}{the lower bound of number of TFs targeting to the gene}
}
\value{
geneTF matrix
}
\description{
Filter geneTF prior based on the gene list user provided
}
