% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_geneTF.R
\name{load_geneTF}
\alias{load_geneTF}
\title{Load TF and target gene matrix prior}
\usage{
load_geneTF(filepath)
}
\arguments{
\item{filepath}{geneTF matrix file path}
}
\value{
geneTF matrix
}
\description{
Load TF and target gene matrix prior
}
